Update ActionTab component
